lina-backend/requirements.txt
langchain
langchain-openai
langserve
fastapi
uvicorn[standard]
python-dotenv
pydantic
sse_starlette
requests
